tasks/scanline/requirements.txt
python-barcode
pillow
